Refact logout logic
